+++ api/pkg/services/postService.go
@@ -9,7 +9,7 @@ import (
 
 type IPostService interface {
 	CreatePost(post *models.Post) error
-	GetFeedPosts(userId int, offset int) ([]*feedPostJSON, error)
+	GetFeedPosts(userId int64, offset int) ([]*feedPostJSON, error)
 }
 
 // Controller contains the service, which contains database-related logic, as an injectable dependency, allowing us to decouple business logic from db logic.
@@ -51,7 +51,7 @@ func (s *PostService) CreatePost(post *models.Post) error {
 	return err
 }
 
-func (s *PostService) GetFeedPosts(userId int, offset int) ([]*feedPostJSON, error) {
+func (s *PostService) GetFeedPosts(userId int64, offset int) ([]*feedPostJSON, error) {
 	// fmt.Println("userId", userId)
 	posts, err := s.PostRepository.GetAllFeedPosts(userId, offset)
